Add exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -31,9 +31,21 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
-  const result = await Blog.findByIdAndDelete(req.params.id)
+  // Decode the token
+  const decodedToken = jwt.verify(req.token, config.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: 'token invalid' })
+  }
 
-  res.status(204).end()
+  // Verify the user
+  const blog = await Blog.findById(req.params.id)
+  const user = await User.findById(decodedToken.id)
+  if (blog.user._id.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(req.params.id)
+    res.status(204).end()
+  } else {
+    res.status(401).json({ error: 'no permission: invalid user' })
+  }
 })
 
 blogsRouter.put('/:id', async (req, res) => {
